Moved the Books model to the books app
